+++ src/writer/mod.rs
@@ -196,7 +196,33 @@ fn write_start_edge(
             write_non_element_node(node, out);
             write_newline(opt.indent, out);
         }
-        NodeType::Declaration |
+        NodeType::Declaration => {
+            depth.write_indent(out);
+            out.extend_from_slice(b"<?xml");
+
+            // Write only possible attributes.
+
+            let attrs = node.attributes();
+
+            if let Some(attr) = attrs.get(AttributeId::Version) {
+                write_attribute(&attr, depth, attrs_depth, opt, out);
+            } else {
+                // The version must always be set.
+                let attr = Attribute::from((AttributeId::Version, "1.0"));
+                write_attribute(&attr, depth, attrs_depth, opt, out);
+            }
+
+            if let Some(attr) = attrs.get("encoding") {
+                write_attribute(&attr, depth, attrs_depth, opt, out);
+            }
+
+            if let Some(attr) = attrs.get("standalone") {
+                write_attribute(&attr, depth, attrs_depth, opt, out);
+            }
+
+            out.extend_from_slice(b"?>");
+            write_newline(opt.indent, out);
+        }
         NodeType::Comment => {
             depth.write_indent(out);
             write_non_element_node(node, out);
@@ -213,9 +239,6 @@ fn write_start_edge(
 /// Specifically: Declaration, Comment, Cdata and Text.
 fn write_non_element_node(node: &Node, out: &mut Vec<u8>) {
     match node.node_type() {
-        NodeType::Declaration => {
-            write_node(b"<?xml ", &node.text(), b"?>", out);
-        }
         NodeType::Comment => {
             write_node(b"<!--", &node.text(), b"-->", out);
         }